libril: Fix rild crash for dial.
The used prebuild vendor ril expects an extra field in the RIL_DIAL
struct. This field needs to be set to an valid address
otherwise rild will crash. Since the meaning of this extra reference is
unkown it is initialized with null.

Change-Id: I37874b8f2c4fcc4654715253eed8677a5257d9ac

Original commit: 3978f585140e9e1250d2765d735458e914eacb08